filter only the owner can delete quotes ij quotesController + test

+++ src/server/controllers/quotesController.ts
@@ -67,16 +67,22 @@ export const deleteQuote = async (
   next: NextFunction
 ) => {
   const { id: quoteId } = req.query;
+  const { user: userId } = req.body;
+  const deleteError = new CustomError(
+    400,
+    "Error deleting quote",
+    "Couldn't delete the quote"
+  );
 
   try {
+    const quoteToDelete = await QuoteModel.findById(quoteId);
+    if (quoteToDelete.owner !== userId) {
+      next(deleteError);
+      return;
+    }
     await QuoteModel.findByIdAndDelete(quoteId);
     res.status(200).json("Quote deleted correctly");
   } catch (error) {
-    const deleteError = new CustomError(
-      400,
-      "Error deleting quote",
-      "Couldn't delete the quote"
-    );
     next(deleteError);
   }
 };
